+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -259,7 +259,7 @@ function Cell(props) {
         isShiftPressed: event.shiftKey,
         isMetaPressed: event.metaKey || event.ctrlKey,
         isCellSelectable: (!isMasked && isSelectable),
-      });
+      }, event);
     }
   };
 
@@ -305,7 +305,7 @@ function Cell(props) {
               isShiftPressed: event.shiftKey,
               isMetaPressed: event.metaKey || event.ctrlKey,
               isCellSelectable: (!isMasked && isSelectable),
-            });
+            }, event);
           }
 
           // Allow default behavior if the event target is an editable field