+++ packages/terra-table/src/subcomponents/Row.jsx
@@ -80,6 +80,7 @@ const propTypes = {
    * Callback function that will be called when a cell in the row is selected.
    * @param {string} rowId rowId
    * @param {string} columnId columnId
+   * @param {object} event event
    */
   onCellSelect: PropTypes.func,